move theme function to utils folder
